15: implement loading chunks, its basic function and a chunk manager, also played a bit with the perlin noise parameter

+++ include/perlin.h
@@ -4,6 +4,6 @@
 
 float perlin2D(float x, float z, float frequency, int seed);
 
-float fractalPerlin2D(float x, float z, float frequency, float octave, float persistence, int seed);
+float fractalPerlin2D(float x, float z, float frequency, int octave, float persistence, int seed);
 
 #endif
